requirements.txt
aiohttp>=3.8.0
asyncio
python-dotenv>=1.0.0
PyGithub>=1.59.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
